mainproject/mykee_retrive.py

import google.generativeai as genai
from langchain_google_genai import GoogleGenerativeAI


model = genai.GenerativeModel('gemini-1.5-flash')


genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))
model = genai.GenerativeModel('gemini-1.5-flash')


retriever = vector_store.as_retriever(
    search_type="similarity_score_threshold",
    search_kwargs={"k": 1, "score_threshold": 0.5},
)
retriever.invoke("Stealing from the bank is a crime", filter={"source": "news"})